requirements.txt
streamlit
streamlit-webrtc
opencv-python
tensorflow
keras
numpy
pandas
imutils
Pillow
streamlit-camera-input-live
